Update single entry stream API docs (#266)
It may not be immediately apparent why accessing an entry by hash uses
an endpoint that ends in /stream. The reason for this is that it may be
useful for a payment processor to watch a hash before the entry has been
posted, waiting for it to get confirmed.

This commit clarifies why entries are streamable, rather than always
downloading instantly like /accounts/{publicKey} does.

Co-authored-by: Waldo Lemmer <>